requirements.txt
piexif==1.1.3
halo==0.0.31
requests==2.31.0
pillow==10.3.0